Destroy object script is now a TriggeredAction
